+++ server/edge/src/protocol.rs
@@ -81,7 +81,7 @@ pub enum ServerMsg {
     /// sip never drags the whole payload. `need` is the packed gameplay row
     /// `value:16 | kind:12 | variant:4` (u16 fixed-point on the need's authored domain);
     /// `set_tic` anchors the lazy eval. Joined by `entity_reference`.
-    Need { entity_reference: u32, zone: u16, need: u32, set_tic: u16 },
+    Need { entity_reference: u32, zone: u16, need: u64, set_tic: u16 },
     /// One `inventory` sub-table row (inventory F2) — a held item in a pawn's slot.
     /// `item = 0` = the slot is now EMPTY (the drop/removal relay — kind-0 suppression's
     /// precedent); `state` is the RESERVED word, echoed for the item-as-entity successor.
